move util to main project
